+++ .gitignore
@@ -9,7 +9,7 @@ __pycache__/
 # Distribution / packaging
 .Python
 env/
-venv/
+venv*/
 build/
 develop-eggs/
 dist/
