--- .travis.yml
@@ -1,6 +1,0 @@
-sudo: false
-language: python
-python:
-    - "3.6"
-install: pip install .
-script: pytest
